+++ docs/source/changelog.rst
@@ -11,6 +11,7 @@ ChangeLog
 ************
 
 * The ``CA_DIGEST_ALGORITHM`` was removed. Use :ref:`settings-ca-default-signature-hash-algorithm` instead.
+* The ``CA_DEFAULT_ECC_CURVE`` was removed. Use :ref:`settings-ca-default-elliptic-curve` instead.
 
 .. _changelog-1.24.0: